+++ lib/l10ngen/app_localizations_de.dart
@@ -1006,7 +1006,7 @@ class AppLocalizationsDe extends AppLocalizations {
   String get editEntryLocationDialogLongitude => 'Längengrad';
 
   @override
-  String get editEntryLocationDialogTimeShift => 'Time shift';
+  String get editEntryLocationDialogTimeShift => 'Zeitverschiebung';
 
   @override
   String get locationPickerUseThisLocationButton => 'Diesen Standort verwenden';
@@ -1018,7 +1018,7 @@ class AppLocalizationsDe extends AppLocalizations {
   String get removeEntryMetadataDialogTitle => 'Entfernung von Metadaten';
 
   @override
-  String get removeEntryMetadataDialogAll => 'All';
+  String get removeEntryMetadataDialogAll => 'Alle';
 
   @override
   String get removeEntryMetadataDialogMore => 'Mehr';
@@ -1497,6 +1497,9 @@ class AppLocalizationsDe extends AppLocalizations {
   String get sortByDuration => 'Nach Dauer';
 
   @override
+  String get sortByPath => 'Nach Pfad';
+
+  @override
   String get sortOrderNewestFirst => 'Neueste zuerst';
 
   @override
@@ -1624,6 +1627,9 @@ class AppLocalizationsDe extends AppLocalizations {
 
   @override
   String get searchDateSectionTitle => 'Datum';
+
+  @override
+  String get searchFormatSectionTitle => 'Formate';
 
   @override
   String get searchAlbumsSectionTitle => 'Alben';