+++ lib/l10ngen/app_localizations_ro.dart
@@ -991,7 +991,7 @@ class AppLocalizationsRo extends AppLocalizations {
   String get editEntryLocationDialogLongitude => 'Longitudine';
 
   @override
-  String get editEntryLocationDialogTimeShift => 'Time shift';
+  String get editEntryLocationDialogTimeShift => 'Schimb de timp';
 
   @override
   String get locationPickerUseThisLocationButton => 'Utilizați această locație';
@@ -1003,7 +1003,7 @@ class AppLocalizationsRo extends AppLocalizations {
   String get removeEntryMetadataDialogTitle => 'Eliminarea metadatelor';
 
   @override
-  String get removeEntryMetadataDialogAll => 'All';
+  String get removeEntryMetadataDialogAll => 'Toate';
 
   @override
   String get removeEntryMetadataDialogMore => 'Mai mult';
@@ -1452,6 +1452,9 @@ class AppLocalizationsRo extends AppLocalizations {
   String get sortByDuration => 'După durată';
 
   @override
+  String get sortByPath => 'După cale';
+
+  @override
   String get sortOrderNewestFirst => 'Cele mai noi mai întâi';
 
   @override
@@ -1579,6 +1582,9 @@ class AppLocalizationsRo extends AppLocalizations {
 
   @override
   String get searchDateSectionTitle => 'Dată';
+
+  @override
+  String get searchFormatSectionTitle => 'Formate';
 
   @override
   String get searchAlbumsSectionTitle => 'Albume';