+++ lib/widgets/viewer/visual/controller_mixin.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'dart:math';
 
 import 'package:aves/app_mode.dart';
+import 'package:aves/model/device.dart';
 import 'package:aves/model/entry/entry.dart';
 import 'package:aves/model/entry/extensions/multipage.dart';
 import 'package:aves/model/entry/extensions/props.dart';
@@ -129,7 +130,7 @@ mixin EntryViewControllerMixin<T extends StatefulWidget> on State<T> {
   }
 
   Future<void> _initVideoController(AvesEntry entry) async {
-    final controller = context.read<VideoConductor>().getOrCreateController(entry);
+    final controller = await context.read<VideoConductor>().getOrCreateController(entry);
     setState(() {});
 
     if (videoAutoPlayEnabled || entry.isAnimated) {
@@ -156,7 +157,9 @@ mixin EntryViewControllerMixin<T extends StatefulWidget> on State<T> {
     if (videoPageEntries.isNotEmpty) {
       // init video controllers for all pages that could need it
       final videoConductor = context.read<VideoConductor>();
-      videoPageEntries.forEach((entry) => videoConductor.getOrCreateController(entry, maxControllerCount: videoPageEntries.length));
+      await Future.forEach(videoPageEntries, (entry) async {
+        await videoConductor.getOrCreateController(entry, maxControllerCount: videoPageEntries.length);
+      });
 
       // auto play/pause when changing page
       Future<void> _onPageChanged() async {
@@ -227,40 +230,40 @@ mixin EntryViewControllerMixin<T extends StatefulWidget> on State<T> {
   static const _pipRatioMin = Rational(18, 43);
 
   Future<void> updatePictureInPicture(BuildContext context) async {
-    if (context.mounted) {
-      if (settings.videoBackgroundMode == VideoBackgroundMode.pip) {
-        final playingController = context.read<VideoConductor>().getPlayingController();
-        if (playingController != null) {
-          final entrySize = playingController.entry.displaySize;
-          final entryAspectRatio = entrySize.aspectRatio;
-          final Rational pipAspectRatio;
-          if (entryAspectRatio > _pipRatioMax.aspectRatio) {
-            pipAspectRatio = _pipRatioMax;
-          } else if (entryAspectRatio < _pipRatioMin.aspectRatio) {
-            pipAspectRatio = _pipRatioMin;
-          } else {
-            pipAspectRatio = Rational(entrySize.width.round(), entrySize.height.round());
-          }
+    if (!device.supportPictureInPicture) return;
 
-          final viewSize = MediaQuery.sizeOf(context) * MediaQuery.devicePixelRatioOf(context);
-          final fittedSize = applyBoxFit(BoxFit.contain, entrySize, viewSize).destination;
-          final sourceRectHint = Rectangle<int>(
-            ((viewSize.width - fittedSize.width) / 2).round(),
-            ((viewSize.height - fittedSize.height) / 2).round(),
-            fittedSize.width.round(),
-            fittedSize.height.round(),
-          );
+    if (context.mounted && settings.videoBackgroundMode == VideoBackgroundMode.pip) {
+      final playingController = context.read<VideoConductor>().getPlayingController();
+      if (playingController != null) {
+        final entrySize = playingController.entry.displaySize;
+        final entryAspectRatio = entrySize.aspectRatio;
+        final Rational pipAspectRatio;
+        if (entryAspectRatio > _pipRatioMax.aspectRatio) {
+          pipAspectRatio = _pipRatioMax;
+        } else if (entryAspectRatio < _pipRatioMin.aspectRatio) {
+          pipAspectRatio = _pipRatioMin;
+        } else {
+          pipAspectRatio = Rational(entrySize.width.round(), entrySize.height.round());
+        }
 
-          try {
-            final status = await Floating().enable(OnLeavePiP(
-              aspectRatio: pipAspectRatio,
-              sourceRectHint: sourceRectHint,
-            ));
-            debugPrint('Enabled picture-in-picture with status=$status');
-            return;
-          } on PlatformException catch (e, stack) {
-            await reportService.recordError(e, stack);
-          }
+        final viewSize = MediaQuery.sizeOf(context) * MediaQuery.devicePixelRatioOf(context);
+        final fittedSize = applyBoxFit(BoxFit.contain, entrySize, viewSize).destination;
+        final sourceRectHint = Rectangle<int>(
+          ((viewSize.width - fittedSize.width) / 2).round(),
+          ((viewSize.height - fittedSize.height) / 2).round(),
+          fittedSize.width.round(),
+          fittedSize.height.round(),
+        );
+
+        try {
+          final status = await Floating().enable(OnLeavePiP(
+            aspectRatio: pipAspectRatio,
+            sourceRectHint: sourceRectHint,
+          ));
+          debugPrint('Enabled picture-in-picture with status=$status');
+          return;
+        } on PlatformException catch (e, stack) {
+          await reportService.recordError(e, stack);
         }
       }
     }